[US_MI] Ingest Supervision Violations (Recidiviz/recidiviz-data#17401)
## Description of the change

Construction of entities:
- In MI, there is a ADH_SUPERVISION_VIOLATION table which holds reports
of supervision violation as well as a ADH_SUPV_VIOLATION_INCIDENT table
which holds data on each incident. Multiple incidents can be attached to
the same record in ADH_SUPERVISION_VIOLATION so for our schema, we're
going to count each incident as a single StateSupervisionViolation and
each record in ADH_SUPERVISION_VIOLATION as a single
StateSupervisionViolationResponse. There are cases where there's a
record in the ADH_SUPERVISION_VIOLATION table but no associated
incidents, and I think these are largely probation violations which MI
research has said we should expect to be less documented in their data.
In those cases, we'll still have a StateSupervisionViolation and
StateSupervisionViolationResponse constructed for them but leave many
fields null.
- Then there's the ADH_MANAGEMENT_DECISION table which holds supervision
officer level decisions/response while the ADH_SANCTION table holds
other sanction responses. Each row in either of these tables attaches to
a record in ADH_SUPERVISION_VIOLATION and so we're going to count each
as a StateSupervisionViolationResponse with an
attached StateSupervisionViolationResponseDecision.
- And finally the ADH_SUPER_COND_VIOLATION tells us about the violated
condition, and each record attaches to a violation incident, which
aligns with our schema.

NOTES:
- Because the ingest view involved `ARRAY_AGG` and `STRUCT`, I feel like
I had to do a finagling with the ingest view parser test expected output
to match what Postgres would output rather than what BQ would output. So
things feel a little janky there...but the input for the
parser/controller tests match the format BQ produces when I run the
ingest view.
- Left some comments with questions in the mappings file, custom enum
parser file, and the utils file

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#16347

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4ba742b13111e338240629c70478de50a90d6dfc