[Workflows] Exclude supervision level unsupervised from all opportunities  (Recidiviz/recidiviz-data#18188)
## Description of the change

In this PR, `dataflow_sessions.correctional_level` = `UNSUPERVISED` is
excluded from all opportunities, as confirmed by Stephanie, and
originally surfaced in Recidiviz/recidiviz-data#17806.

The `compartment_level_2` values of `ABSCONSION`, `BENCH_WARRANT` and
`INTERNAL_UNKNOWN` are also now excluded for
`supervision_level_downgrade` for Idaho.

Additionally, this prompted a renaming of our candidate populations and
a refactoring of the
`supervision_population_eligible_levels_additional_filters` to include a
`correctional_level_filters` argument as well as a
`compartment_level_2_filters` argument.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b2b826fb9aa07bdffc488fc1bcbe4625b2e341ce