Adding encoding flag to `upload_local_file_to_bq` script to allow for non utf-8 uploads (Recidiviz/recidiviz-data#18295)
## Description of the change

When testing a MO encoding issue via the `upload_local_file_to_bq`
script, the encoding being used was cp1252 instead of utf-8. When trying
to specify the encoding, the line in the script before getting the
entire df looked for columns, but didn't specify the different encoding
and therefore was defaulting to utf-8 and causing errors. This change
adds the encoding specification in the first read_csv call so that the
encoding can be applied throughout the entire script.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a4b73774d9ee7aa59ba713c5bba1c82be2b5f8f0