[Workflows] Handle officers with multiple external_ids (Recidiviz/recidiviz-data#18248)
## Description of the change

When we create the new `Roster` users, we will try to include as many
external ids as possible since the `Roster` records will be synced to BQ
to be used by other products like workflows `staff_record`. There were
several users in the system who had multiple external_ids assigned to
them. Some of these have multiple external_ids assigned in the raw data,
and some ended up with multiple ids in workflows `staff_record` due to
misspellings in the names/email addresses in `us_tn_roster`.

I cleaned up `us_tn_roster` and TN no has no users with multiple ids. I
only did this cleanup in production.

We decided to merge the extra ids for other users into a single id when
creating the `Roster` users when launching the admin panel, and map the
staff/client records that belong to the old id to the new id. This is
done in `agent_multiple_ids_map` table. This has been run in
jovergaag_17763

Let me know if you have other ideas how this could be done, or done
better. The document listing the users with multiple ids, and how they
were each handled is
[here](https://docs.google.com/spreadsheets/d/1KprogKRnohzK_3yYuszvru5CnUKbn-696ZXr9PZ8ZEA/edit#gid=685986821).

Once this PR is merged and deployed, we will have a single ID used by
all users in the system which can be added to the roster table.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#17763

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7db7597a898cba7e78770795bb00a500ee2fc7d0