[Justice Counts] Use our DB to manage user-agency relationships instead of Auth0 (Recidiviz/recidiviz-data#18217)
## Description of the change

Currently, the way we decide which agency a user belongs to is pretty
complicated:
* For Recidiviz Admins, we ignore Auth0 and our own DB, and instead give
them permission to *all* agencies. This happens
[here](https://github.com/Recidiviz/recidiviz-data/blob/35d831b7052d0fa3d30f3ba969a16a1b0df8b4b7/recidiviz/justice_counts/control_panel/utils.py#LL168C16-L168C16).
This can result in surprising behavior, e.g. a user has access to an
agency but doesn't appear in that agency's Team Management list
(@mxosman noticed this "ghost behavior" during playtesting).
* For non-Recidiviz Admins, we use Auth0 (agency_ids stored in
app_metadata). This happens
[here](https://github.com/Recidiviz/recidiviz-data/blob/35d831b7052d0fa3d30f3ba969a16a1b0df8b4b7/recidiviz/justice_counts/control_panel/utils.py#L66).
* To decide who appears on the Team Management page for an agency, we
use our own DB (the AgencyUserAccountAssociation table), which
essentially caches the info from Auth0 and makes it so that we don't
have to query Auth0 every time we load this page.

There were a few reasons we decided on this setup:
* For the sake of testing and debugging, we wanted Recidiviz Admins to
have access to all agencies. It was a big pain to manually add every
admin to every new agency we onboarded. So that's why we introduced the
special override behavior.
* We thought it was safer / more secure to have Auth0 be the source of
truth of who had access to what agency.

After thinking about this more deeply, I think neither reason justifies
the complexity of the current situation. I propose instead that for
every user, **regardless of whether or not they are an admin, the
AgencyUserAccountAssociation determines which agency they belong to, and
what their role is at that agency.** We will no longer use Auth0 to
store agency membership or roles. This PR implements that behavior.

If we still want Recidiviz Admins to have access to all agencies, we can
add some sort of hook that does this whenever a new agency is created.
But I suggest we think about that separately.

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#17936

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4c504ee138f488b6b7b5f219d61869f15dd0a478