[Justice Counts] Fix local development issue introduced by Recidiviz/recidiviz-data#18217  (Recidiviz/recidiviz-data#18325)
## Description of the change

In Recidiviz/recidiviz-data#18325 I changed the logic that determines which agencies a user
belongs to. Previously, any admins were just "assumed" to belong to all
agencies. Now, you only belong to an agency if there is a corresponding
row in the AgencyUserAccountAssociation table. This should make things
easier to reason about. However, it broke local development -- now you
would see warnings about not belonging to any agency. It's hard to
ensure that everyone's local database has an up-to-date
AgencyUserAccountAssociation table via fixtures. However, by adding this
little hack in /users PUT, we can ensure that the table gets updated the
first time the user runs locally.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7a5d944f8b5340ee15e9dd1fd0a9008d37aa6cc0