[US_MI] Add 0 as value for response decision enum mapping (Recidiviz/recidiviz-data#18320)
## Description of the change

There's currently an enum parsing error in staging (Recidiviz/recidiviz-data#18294). This PR
adds in the missing value ("0") to the mapping.

Note: technically, `_decision='0'` is the reference code for "TEMP", and
it seems like there are quite a few (7%) management decisions that are
designated with this decision type. Not sure if this just means the
decision hasn't been made yet, but the decision date is valued for these
records. In the data, we'll usually see another non-"0" decision made
after this "0" decision, but not necessarily by the same person. I'm
gonna keep this in the ingest view results for now and it can be
filtered out later as necessary.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5ed42a3633150a8ab05f4137c70b97907c15c598