[US_IX] Bump machine size for fuzzy matching pipeline (Recidiviz/recidiviz-data#18317)
## Description of the change

The IX case notes pipeline is flaky, failing with worker timeouts pretty
frequently. We think this is likely due to hotspotting, where one person
has a lot of case notes and it overloads a particular worker.

For now, I'm just increasing to the next machine size to see if that
fixes the issue and buys us some time.

Long term, we can change the pipeline to just group by case note instead
of person, which should give us a better distribution of data size
across keys. There are also a couple other solutions we could try here:
https://cloud.google.com/dataflow/docs/guides/common-errors#worker-lost-contact

It also looks like we could have dataflow do vertical scaling if we
enabled Dataflow Prime:
https://cloud.google.com/dataflow/docs/guides/enable-dataflow-prime

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#18160

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f56e4f0c9bd5f681a661070df6ac349d3bbc5821