[Airflow] Wait for normalized_state update to complete before starting metric pipelines (Recidiviz/recidiviz-data#18246)
## Description of the change

Currently, we are refreshing the `normalized_state` dataset *at the same
time* as the metric pipelines (after the normalization pipeline runs).
Since the metric pipelines now pull data from the `normalized_state`
dataset, we need to make sure we're waiting for the dataset refresh to
complete.

The normalized_state update takes about 7 min so this will increase the
end-to-end runtime of our full orchestration DAG by about 7 min.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#18243.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0eaf67bb46d00ff316373c06f8b348bfa17ca9bf