[US_PA] Change "true" to "True" in CCIS mapping for evaluating start_is_new_act_122_admission (Recidiviz/recidiviz-data#18345)
## Description of the change

Recidiviz/recidiviz-data#18009 updated the CCIS mapping such that incarceration periods with
`start_is_new_act_122_admission=true` are given an admission_reason of
INTERNAL_UNKNOWN, which should avoid periods with a PFI of
INTERNAL_UNKNOWN and an admission_reason of REVOCATION. However,
failures in the invalid_admission_reason_and_pfi validation indicated
that this mapping was not functioning properly. This was found to be
related to the start_is_new_act_122_admission column's case: changing
the condition to `start_is_new_act_122_admission=True` fixes this error.

This change will require a partial rerun to take effect.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#17074

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e0ecb0bd78ed5f8afd50df05b7c2160d9165c39e