[Justice Counts] add role to spreadsheet endpoint (Recidiviz/recidiviz-data#18318)
## Description of the change

This adds the role to the `"uploaded_by"` field in the spreadsheet
endpoint. This will **not** break the FE because there is a v2 field, so
we will not need to land this with the FE changes for this endpoint.

## Related issues

Associated with Recidiviz/recidiviz-data#17967

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ba7f47f80bffb2ada2fdf14b98a3ed0bbaab44ed